Admin review delete

See merge request s02-bigdata-sub3/s02p23d106!38

Former-commit-id: 4344069d6b85cb7b9c77808f23680ac15013c6e0 [formerly a7354749ef0a3059f5741be05bcca1c5b1da07f1]
Former-commit-id: 39d2a1e64b1ded9c27de0b371e191488eb6f696c

+++ backend/api/views.py
@@ -286,25 +286,31 @@ class StoreReviewSet(viewsets.ModelViewSet):
 
     def create(self, request):
         # 리뷰를 작성하는 함수입니다.
-        # if request.user.is_authenticated:
-        #     data = request.data
-        #     review = models.Review.objects.create(store_id=data["store"], user_id=request.user.id, content=data["content"], score=data["score"], reg_time=datetime.datetime.now())
-        #     return Response("작성 성공")
-        # else:
-        #     return Response("작성 실패")
-        data = request.data
-        store = Store.objects.get(id=data["store"])
-        store_name = store.store_name
-        user = CustomUser.objects.get(id=data["user"])
-
-        # 받아온 데이터를 이용해서 Review를 작성합니다.
-        models.Review.objects.create(store_id=data["store"], user_id=data["user"], content=data["content"], score=data["score"], reg_time=datetime.datetime.now(), store_name=store_name)
+        print('asdf')
+        if request.user.is_authenticated:
+            data = request.data
+            store = Store.objects.get(id=data["store"])
+            user = request.user
+            review = models.Review.objects.create(store=store, user=user, content=data["content"], score=data["score"])
+            store.review_count += 1
+            store.save()
+            user.review_count += 1
+            user.save()
+            return Response("작성 성공")
+        else:
+            return Response("작성 실패")
+        # data = request.data
+        # store = Store.objects.get(id=data["store"])
+        # store_name = store.store_name
+        # user = CustomUser.objects.get(id=data["user"])
+        # # 받아온 데이터를 이용해서 Review를 작성합니다.
+        # models.Review.objects.create(store_id=data["store"], user_id=data["user"], content=data["content"], score=data["score"], reg_time=datetime.datetime.now(), store_name=store_name)
         
         # 작성이 완료되었다면 매장과 유저의 review_count를 1씩 추가합니다.
-        store.review_count += 1
-        store.save()
-        user.review_count += 1
-        user.save()
+        # store.review_count += 1
+        # store.save()
+        # user.review_count += 1
+        # user.save()
         return Response("작성 성공")
 
     def update(self, request, pk=None):
@@ -323,15 +329,28 @@ class StoreReviewSet(viewsets.ModelViewSet):
         '''
         받아온 pk에 해당하는 리뷰를 삭제합니다.
         '''
-        review = models.Review.objects.get(id=pk)
-        user = models.CustomUser.objects.get(id=review.user_id)
-        store = models.Store.objects.get(id=review.store_id)
-        review.delete()
-        user.review_count -= 1
-        user.save()
-        store.review_count -= 1
-        store.save()
-        return Response("삭제 성공")
+        if request.user.is_authenticated:
+            review = models.Review.objects.get(id=pk)
+            user = review.user
+            store = review.store
+            if user.id == review.user_id:
+                review.delete()
+                user.review_count -= 1
+                user.save()
+                store.review_count -= 1
+                store.save()
+                return Response("삭제 성공")
+        elif request.user.is_staff:
+            review = models.Review.objects.get(id=pk)
+            user = review.user
+            store = review.store
+            review.delete()
+            user.review_count -= 1
+            user.save()
+            store.review_count -= 1
+            store.save()
+            return Response("삭제 성공")
+        return Response("삭제 실패")
 
 
 @api_view(['GET'])
